scripts/train.py
# scripts/train.py
import hydra
from omegaconf import DictConfig, OmegaConf
import logging

logger = logging.getLogger(__name__)

@hydra.main(version_base=None, config_path="../configs", config_name="config")
def main(cfg: DictConfig):
    logger.info("Initializing SC-MAC Research Pipeline...")
    logger.info(f"Configuration:\n{OmegaConf.to_yaml(cfg)}")
    
    # Placeholder for Phase 2: Agent Initialization
    # agent = instantiate(cfg.agent)
    
    print("✅ Project Setup Complete.")

if __name__ == "__main__":
    main()