requirements.txt
# Core ML
torch>=2.4.0
transformers>=4.44.0
accelerate>=0.33.0
hydra-core>=1.3.2
pytorch-lightning>=2.4.0

# Multi-Agent & Federated
ray>=2.34.0
flwr>=1.10.0  # Flower for Federated Learning

# Logic & Evaluation
z3-solver>=4.13.0  # For Symbolic Logic Checking
sentencepiece
wandb  # Experiment tracking
pytest